[CITE-85] - Addressed review comments - moved async response classes to appropriate package, fixed delete confirmation text, removed tests for AsyncCitationManager
