+++ citesphere/src/main/webapp/WEB-INF/views/auth/group/items.html
@@ -522,7 +522,7 @@ function pollStatus(){
         <h4 class="modal-title">Delete Entry</h4>
       </div>
       <div class="modal-body">
-        <p>Are you sure you want to delete the selected Entry</p>
+        <p>Are you sure you want to delete the selected Entries?</p>
       </div>
       <div class="modal-footer">
         <button type="button" class="btn btn-default" data-dismiss="modal">No, cancel!</button>
